requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
distlib==0.3.9
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
h5py==3.13.0
identify==2.6.10
idna==3.10
iniconfig==2.1.0
invoke==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
protobuf==6.30.2
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
ruff==0.11.8
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tomli==2.2.1
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.1
wandb==0.19.10
wcwidth==0.2.13
websocket-client==1.8.0
